Makes a Property aware of the Mapper in charge to generate and parse the expected database format.
