+++ src/main/java/sirius/db/jdbc/batch/InsertQuery.java
@@ -63,7 +63,7 @@ public class InsertQuery<E extends SQLEntity> extends BatchQuery<E> {
             PreparedStatement stmt = prepareStmt();
             int i = 1;
             for (Property property : getProperties()) {
-                stmt.setObject(i++, property.getValueForDatasource(entity));
+                stmt.setObject(i++, property.getValueForDatasource(OMA.class, entity));
             }
 
             if (invokeChecks) {
