+++ src/main/java/sirius/db/mixing/properties/LocalDateProperty.java
@@ -6,23 +6,29 @@
  * http://www.scireum.de - info@scireum.de
  */
 
-package sirius.db.es.properties;
+package sirius.db.mixing.properties;
 
 import com.alibaba.fastjson.JSONObject;
 import sirius.db.es.ESPropertyInfo;
-import sirius.db.es.ElasticEntity;
 import sirius.db.es.IndexMappings;
 import sirius.db.es.annotations.ESOption;
 import sirius.db.es.annotations.IndexMode;
+import sirius.db.jdbc.schema.SQLPropertyInfo;
+import sirius.db.jdbc.schema.Table;
+import sirius.db.jdbc.schema.TableColumn;
 import sirius.db.mixing.AccessPath;
 import sirius.db.mixing.EntityDescriptor;
 import sirius.db.mixing.Property;
 import sirius.db.mixing.PropertyFactory;
+import sirius.db.mongo.QueryBuilder;
 import sirius.kernel.commons.Strings;
 import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Register;
+import sirius.kernel.nls.NLS;
 
 import java.lang.reflect.Field;
+import java.sql.Date;
+import java.sql.Types;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.function.Consumer;
@@ -31,7 +37,7 @@ import java.util.function.Consumer;
  * Represents a date property which contains no associated time value. This is used to represents fields of type
  * {@link LocalDate}.
  */
-public class ESLocalDateProperty extends Property implements ESPropertyInfo {
+public class LocalDateProperty extends Property implements ESPropertyInfo, SQLPropertyInfo {
 
     /**
      * Factory for generating properties based on their field type
@@ -41,7 +47,7 @@ public class ESLocalDateProperty extends Property implements ESPropertyInfo {
 
         @Override
         public boolean accepts(EntityDescriptor descriptor, Field field) {
-            return ElasticEntity.class.isAssignableFrom(descriptor.getType()) && LocalDate.class.equals(field.getType());
+            return LocalDate.class.equals(field.getType());
         }
 
         @Override
@@ -49,21 +55,30 @@ public class ESLocalDateProperty extends Property implements ESPropertyInfo {
                            AccessPath accessPath,
                            Field field,
                            Consumer<Property> propertyConsumer) {
-            propertyConsumer.accept(new ESLocalDateProperty(descriptor, accessPath, field));
+            propertyConsumer.accept(new LocalDateProperty(descriptor, accessPath, field));
         }
     }
 
-    protected ESLocalDateProperty(EntityDescriptor descriptor, AccessPath accessPath, Field field) {
+    protected LocalDateProperty(EntityDescriptor descriptor, AccessPath accessPath, Field field) {
         super(descriptor, accessPath, field);
     }
 
     @Override
     public Object transformValue(Value value) {
-        return value.asLocalDate(null);
+        return NLS.parseUserString(LocalDate.class, value.asString());
     }
 
     @Override
-    protected Object transformFromDatasource(Value object) {
+    protected Object transformFromJDBC(Value data) {
+        Object object = data.get();
+        if (object == null) {
+            return null;
+        }
+        return ((Date) object).toLocalDate();
+    }
+
+    @Override
+    protected Object transformFromElastic(Value object) {
         String valueAsString = object.asString();
         if (Strings.isEmpty(valueAsString)) {
             return null;
@@ -77,12 +92,36 @@ public class ESLocalDateProperty extends Property implements ESPropertyInfo {
     }
 
     @Override
-    protected Object transformToDatasource(Object object) {
+    protected Object transformFromMongo(Value object) {
+        return object.asLocalDate(null);
+    }
+
+    @Override
+    protected Object transformToJDBC(Object object) {
+        return object == null ? null : Date.valueOf((LocalDate) object);
+    }
+
+    @Override
+    protected Object transformToElastic(Object object) {
         if (!(object instanceof LocalDate)) {
             return null;
         }
 
         return DateTimeFormatter.ISO_LOCAL_DATE.format((LocalDate) object);
+    }
+
+    @Override
+    protected Object transformToMongo(Object object) {
+        if (!(object instanceof LocalDate)) {
+            return null;
+        }
+
+        return QueryBuilder.FILTERS.transform(object);
+    }
+
+    @Override
+    public void contributeToTable(Table table) {
+        table.getColumns().add(new TableColumn(this, Types.DATE));
     }
 
     @Override
