+++ src/main/java/sirius/db/mixing/properties/LocalDateTimeProperty.java
@@ -6,23 +6,29 @@
  * http://www.scireum.de - info@scireum.de
  */
 
-package sirius.db.es.properties;
+package sirius.db.mixing.properties;
 
 import com.alibaba.fastjson.JSONObject;
 import sirius.db.es.ESPropertyInfo;
-import sirius.db.es.ElasticEntity;
 import sirius.db.es.IndexMappings;
 import sirius.db.es.annotations.ESOption;
 import sirius.db.es.annotations.IndexMode;
+import sirius.db.jdbc.Databases;
+import sirius.db.jdbc.schema.SQLPropertyInfo;
+import sirius.db.jdbc.schema.Table;
+import sirius.db.jdbc.schema.TableColumn;
 import sirius.db.mixing.AccessPath;
 import sirius.db.mixing.EntityDescriptor;
 import sirius.db.mixing.Property;
 import sirius.db.mixing.PropertyFactory;
+import sirius.db.mongo.QueryBuilder;
 import sirius.kernel.commons.Strings;
 import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Register;
+import sirius.kernel.nls.NLS;
 
 import java.lang.reflect.Field;
+import java.sql.Types;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
@@ -32,7 +38,7 @@ import java.util.function.Consumer;
  * Represents a timestamp property which contains a date along with a time value. This is used to represents fields of
  * type {@link LocalDateTime}.
  */
-public class ESLocalDateTimeProperty extends Property implements ESPropertyInfo {
+public class LocalDateTimeProperty extends Property implements ESPropertyInfo, SQLPropertyInfo {
 
     /**
      * Factory for generating properties based on their field type
@@ -42,8 +48,7 @@ public class ESLocalDateTimeProperty extends Property implements ESPropertyInfo 
 
         @Override
         public boolean accepts(EntityDescriptor descriptor, Field field) {
-            return ElasticEntity.class.isAssignableFrom(descriptor.getType())
-                   && LocalDateTime.class.equals(field.getType());
+            return LocalDateTime.class.equals(field.getType());
         }
 
         @Override
@@ -51,21 +56,30 @@ public class ESLocalDateTimeProperty extends Property implements ESPropertyInfo 
                            AccessPath accessPath,
                            Field field,
                            Consumer<Property> propertyConsumer) {
-            propertyConsumer.accept(new ESLocalDateTimeProperty(descriptor, accessPath, field));
+            propertyConsumer.accept(new LocalDateTimeProperty(descriptor, accessPath, field));
         }
     }
 
-    protected ESLocalDateTimeProperty(EntityDescriptor descriptor, AccessPath accessPath, Field field) {
+    protected LocalDateTimeProperty(EntityDescriptor descriptor, AccessPath accessPath, Field field) {
         super(descriptor, accessPath, field);
     }
 
     @Override
     public Object transformValue(Value value) {
-        return value.asLocalDateTime(null);
+        return NLS.parseUserString(LocalDateTime.class, value.asString());
     }
 
     @Override
-    protected Object transformFromDatasource(Value object) {
+    protected Object transformFromJDBC(Value object) {
+        Object data = object.get();
+        if (data == null) {
+            return null;
+        }
+        return Databases.decodeLocalDateTime((long) data);
+    }
+
+    @Override
+    protected Object transformFromElastic(Value object) {
         String valueAsString = object.asString();
         if (Strings.isEmpty(valueAsString)) {
             return null;
@@ -79,7 +93,20 @@ public class ESLocalDateTimeProperty extends Property implements ESPropertyInfo 
     }
 
     @Override
-    protected Object transformToDatasource(Object object) {
+    protected Object transformFromMongo(Value object) {
+        return object.asLocalDateTime(null);
+    }
+
+    @Override
+    protected Object transformToJDBC(Object object) {
+        if (object == null) {
+            return null;
+        }
+        return Databases.encodeLocalDateTime((LocalDateTime) object);
+    }
+
+    @Override
+    protected Object transformToElastic(Object object) {
         if (!(object instanceof LocalDateTime)) {
             return null;
         }
@@ -88,10 +115,24 @@ public class ESLocalDateTimeProperty extends Property implements ESPropertyInfo 
     }
 
     @Override
+    protected Object transformToMongo(Object object) {
+        if (!(object instanceof LocalDateTime)) {
+            return null;
+        }
+
+        return QueryBuilder.FILTERS.transform(object);
+    }
+
+    @Override
     public void describeProperty(JSONObject description) {
         description.put(IndexMappings.MAPPING_TYPE, "date");
         transferOption(IndexMappings.MAPPING_STORED, IndexMode::stored, ESOption.ES_DEFAULT, description);
         transferOption(IndexMappings.MAPPING_INDEXED, IndexMode::indexed, ESOption.ES_DEFAULT, description);
         transferOption(IndexMappings.MAPPING_DOC_VALUES, IndexMode::indexed, ESOption.ES_DEFAULT, description);
+    }
+
+    @Override
+    public void contributeToTable(Table table) {
+        table.getColumns().add(new TableColumn(this, Types.BIGINT));
     }
 }
