+++ src/main/java/sirius/db/mixing/properties/StringMapProperty.java
@@ -6,21 +6,24 @@
  * http://www.scireum.de - info@scireum.de
  */
 
-package sirius.db.es.properties;
+package sirius.db.mixing.properties;
 
 import com.alibaba.fastjson.JSONObject;
+import org.bson.Document;
 import sirius.db.es.ESPropertyInfo;
+import sirius.db.es.Elastic;
 import sirius.db.es.ElasticEntity;
 import sirius.db.es.IndexMappings;
 import sirius.db.es.annotations.ESOption;
 import sirius.db.es.annotations.IndexMode;
 import sirius.db.mixing.AccessPath;
+import sirius.db.mixing.BaseMapper;
 import sirius.db.mixing.EntityDescriptor;
 import sirius.db.mixing.Mixing;
 import sirius.db.mixing.Property;
 import sirius.db.mixing.PropertyFactory;
-import sirius.db.mixing.properties.BaseMapProperty;
 import sirius.db.mixing.types.StringMap;
+import sirius.db.mongo.Mango;
 import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Register;
 
@@ -34,14 +37,14 @@ import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 /**
- * Represents an {@link StringMap} field within an {@link ElasticEntity}.
+ * Represents an {@link StringMap} field within an {@link ElasticEntity} or a {@link sirius.db.mongo.MongoEntity}.
  * <p>
  * Note that maps are stored as a list of nested objects which contain a <tt>key</tt> and
  * a <tt>value</tt>. This is used to prevent a mapping explosion within ES while still permitting to
  * search and filter using nested queries. Use {@link IndexMode#indexed()} (set to FALSE) to
  * set the type to <tt>object</tt> - which will still create key/value pairs but stored as objects.
  */
-public class ESStringMapProperty extends BaseMapProperty implements ESPropertyInfo {
+public class StringMapProperty extends BaseMapProperty implements ESPropertyInfo {
 
     /**
      * Contains the name of the field used to store the map key
@@ -61,8 +64,7 @@ public class ESStringMapProperty extends BaseMapProperty implements ESPropertyIn
 
         @Override
         public boolean accepts(EntityDescriptor descriptor, Field field) {
-            return ElasticEntity.class.isAssignableFrom(descriptor.getType())
-                   && StringMap.class.equals(field.getType());
+            return StringMap.class.equals(field.getType());
         }
 
         @Override
@@ -76,33 +78,49 @@ public class ESStringMapProperty extends BaseMapProperty implements ESPropertyIn
                                 field.getDeclaringClass().getName());
             }
 
-            propertyConsumer.accept(new ESStringMapProperty(descriptor, accessPath, field));
+            propertyConsumer.accept(new StringMapProperty(descriptor, accessPath, field));
         }
     }
 
-    ESStringMapProperty(EntityDescriptor descriptor, AccessPath accessPath, Field field) {
+    StringMapProperty(EntityDescriptor descriptor, AccessPath accessPath, Field field) {
         super(descriptor, accessPath, field);
     }
 
-    @SuppressWarnings("unchecked")
     @Override
-    protected Object transformFromDatasource(Value object) {
+    @SuppressWarnings("unchecked")
+    protected Object transformFromElastic(Value object) {
         Map<Object, Object> result = new HashMap<>();
         Object value = object.get();
         if (value instanceof Collection) {
             ((Collection<Map<?, ?>>) value).forEach(entry -> result.put(entry.get(KEY), entry.get(VALUE)));
         }
-
         return result;
     }
 
-    @SuppressWarnings("unchecked")
     @Override
-    protected Object transformToDatasource(Object object) {
+    protected Object transformFromMongo(Value object) {
+        return object.get();
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    protected Object transformToElastic(Object object) {
         return ((Map<?, ?>) object).entrySet()
                                    .stream()
                                    .map(e -> new JSONObject().fluentPut(KEY, e.getKey()).fluentPut(VALUE, e.getValue()))
                                    .collect(Collectors.toList());
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    protected Object transformToMongo(Object object) {
+        if (object instanceof Document) {
+            return object;
+        }
+
+        Document doc = new Document();
+        doc.putAll((Map<String, String>) object);
+        return doc;
     }
 
     @Override
